ci: add lighthouse ci
Add/update github action and npm script for LHCI.
Some assertions are temporarly disabled in order to ease the
transition, but all of them should be handled later.
Except for http2, the lhci server not allowing to enable it when using
a static dir. The audit is run on master *after* the release for the
same reason.