refactor(panpress-cli)!: use an "app" source dir
Using the project root as the base directory for the whole web app leads
to a too complexe and verbose snowpack config.

BREAKING CHANGE: views are now generated in the /app/views/ directory,
and routes.js in /app/js/routes.js. But source files (md) for pages
should stay in the /pages directory.