fix(website): update shell lang on navigation
The router wasn't notifing the shell to update its language when
navigating to a path specify a different language (i.e. /fr/ or
/en/) than the previously selected one. This caused an inconsistency
between the language used for content and for the app shell, especially
when navigating through the history (i.e. "back" button) after the
preferred language was changed.